refactor: split layer core crate
Move the shared hook manager and hook macros out of
`mirrord-layer` into a new `mirrord-layer-core` crate.

This keeps the shared hooking APIs reusable while keeping the Frida
dependency scoped to Unix targets, including the macOS layer builds
that use `HookManager`.
