+++ mirrord/layer/core/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "mirrord-layer-core"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+publish.workspace = true
+edition.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+mirrord-layer-lib = { path = "../../layer-lib" }
+libc.workspace = true
+tracing.workspace = true
+
+[target.'cfg(unix)'.dependencies]
+frida-gum.workspace = true
