+++ mirrord/layer/src/exec_utils.rs
@@ -14,6 +14,7 @@ use mirrord_layer_lib::{
         Detour::{self, Bypass, Error, Success},
     },
     error::HookError,
+    graceful_exit,
 };
 use mirrord_layer_macro::{hook_fn, hook_guard_fn};
 use mirrord_sip::{MIRRORD_PATCH_DIR, SipError, SipPatchOptions, sip_patch};
@@ -24,7 +25,6 @@ use crate::{
     EXECUTABLE_ARGS,
     common::{CheckedInto, strip_mirrord_path},
     exec_hooks::{hooks, *},
-    graceful_exit,
     hooks::HookManager,
     replace,
 };
