+++ mirrord/layer/src/lib.rs
@@ -119,6 +119,7 @@ use nix::{
 };
 use socket::SOCKETS;
 
+pub(crate) use crate::macros::*;
 use crate::{
     common::make_proxy_request_with_response, load::LoadType, socket::hooks::MANAGED_ADDRINFO,
 };
